Make sure " +" is not rendered as a markup
In Asciidoctor, " +" at the end of line forces a line break. Since " + "
is used in several texts, this could happen by accident.

Defining an attribute `{plus}` to replace "+" made no difference:
Somehow it behaves like "+". So `{nbsp}+` is used, i.e. replacing the
previous space with the default attribute of a non-breaking space.